+++ projects/miopen/src/scalar.cpp
@@ -32,7 +32,7 @@ Scalar::Scalar(ConstData_t ptr, miopenDataType_t type) : mType(type)
 {
     if(type == miopenFloat)
     {
-        mVal = *static_cast<const float*>(ptr);
+        mVal = double(*static_cast<const float*>(ptr));
     }
     else if(type == miopenDouble)
     {
